重新分离Engine与Task，主页面已经可以使用[20211005]
Former-commit-id: 079b7dbd387e892fd76e594acd0bfb55f68d6f55

+++ app/src/main/java/com/funny/translation/translation/BasicTranslationTask.kt
@@ -6,7 +6,7 @@ import com.funny.translation.utils.StringUtil
 import java.lang.StringBuilder
 
 
-abstract class BasicTranslationTask(sourceString: String,sourceLanguage : Int, targetLanguage : Int) :
+abstract class BasicTranslationTask(sourceString: String,sourceLanguage : Short, targetLanguage : Short) :
     CoreTranslationTask(sourceString, sourceLanguage, targetLanguage) , TranslationEngine {
 
     @Throws(TranslationException::class)
